requirements.txt
dash
pandas
dash-bootstrap-components
python-dotenv
requests
dash-mantine-components 
dash-iconify
python-dotenv